api/requirements.txt
Pillow
requests
imageio[ffmpeg]
